[project @ 2005-02-07 00:04:26 by duncan_coutts] Various win32 fixes.
configure.ac: Do not define WIN32 on windows since it is already defined on
that platform.
Define a conditional so that Makefile.am can conditonally compile a few modules.
We no longer need to pass -fnative-struct to gcc on windows.

Makefile.am: Add $(EXEEXT) in a few places to so names of dependencies are
correct on windows.
Only compile the plug and socket modules on non-win32 platforms.

C2HSConfig.hs.in: comment out the @TMPDIR@ substitution since it causes
problems in windows and is not used anyway.

CParser.hs: make c2hs accept __attribute__'s on declerations of functions
returning pointer types. This is needed for some mingw header file (float.h).