[project @ 2005-02-14 02:10:47 by duncan_coutts] Lots of little improvements to the generated documentation.
Generalise the "KnownTypes" table to "KnownSymbols" and use that in various
places to make the recognition of things we should hyperlink more accurate.
Convert enum value names to Haskell style names.
Recognise type names and enum value names even when they are not marked up in
the original docs.
Deal with deprecated objects and methods, but only generate deprecated methods
where they exist in the original modules (ie don't add new deprecated methods)
Handle "shared" methods but ignore the *_get_type() ones.
Change to using FM's for the KnownSymbols map so it's not all quite so slow.