Change the memory management for TreeIters.
Instead of using ForeignPtrs to C malloc()ed memory, copy the TreeIter data to
a Haskell structure. TreeIters are small concrete C structs so it'll probably
perform better to just copy them rather than manage them with the C heap and a
ForeignPtr with a finaliser.

Also this means that we now have just one way of managing TreeIters where as
previously we had two: the original by-ref style and a by-value style for the
new CustomStore interface (where we need access to the TreeIter
representation). We now use the by-value style everywhere.

They used to be represented like so:
{#pointer *TreeIter foreign newtype#}

We now use:
data TreeIter = TreeIter {-# UNPACK #-} !CInt !Word !Word !Word
{#pointer *TreeIter as TreeIterPtr -> TreeIter #}
instance Storable TreeIter where ...