Repair general DND functionality on lists.
This patch not only represents some bug fixes, it also adds a few tags that
make the actual use of DND easier. Some functions such as selectionDataGet
have changed in order to make them safer. It should be easy now to add
Clipboards. What needs to be tested is if all functions are there to
implement DND on normal widgets. An example for this would be quite
elaborate, hence I have no example yet.