Makefile.am: add a "dep" target that will build .dep files
this target only builds the dependencies that are needed early in the
build process (namely, tools/c2hs/c2hsLocal.deps and the files
generated by chsDepend).  It is depended on by the "all" target, and
so will be processed before anything else, since all recursively runs
make all-am, which does the actual building.