Allow for NULL object to be passed.
This patch not only allows for NULL objects but changes the way Objects that are passed to signal handlers are wrapped. They are now all unref'ed using the main loop method which is correct for Gtk and everything that builds on Gtk. If there ever is a module that has singals but does not depend on Gtk, then this approach has to change. It seems that the template and hierarchy file should be moved to Gtk and that the unref function should become an argument to the callback generator.