make sure ListStore implements TreeModelIface correctly
The current implementation of the TreeModelIface interface by ListStore
is not 100% correct. This causes a problem with combo boxes not being
rendered correctly when they don't have an Entry box. This only occurs
under GTK 3.

The main issue is with treeModelIfaceIterChildren: this function is
sometimes called with Nothing as its argument by GTK, even for a flat
list store. That happens when GTK is trying to obtain an iterator
pointing to the first element of the list. The current implementation
returns unconditionally Nothing, which is wrong.

There is also a similar issue with treeModelIfaceGetIter which can
return an out-of-bounds iterator and causes GTK 3 to display an
additional blank element in combo boxes.