Fix fpco/stackage#1523
It looks like haddock does not like some of the comments in the code
I moved into the gtk2hs-buildtools library.  The simplest fix seems to
be to add an extra space to `-- *` (to make it `--  *`) so that haddock
does not try to process it.

I was able to reproduce the issue locally, however I could not
reproduce it on travis.